requirements.txt
python-telegram-bot --pre
python-telegram-bot[job-queue]
pandas